Editorial changes to pr 59; merge Mux2/Mux4.
