+++ src/main/scala/problems/Max2.scala
@@ -5,8 +5,8 @@ import chisel3._
 
 // Problem:
 //
-// Implement test for this module. Please edit:
-// $chisel_tutorial/src/test/scala/problems/Max2.scala
+// Implement a test for this module. Please edit:
+// .../chisel_tutorial/src/test/scala/problems/Max2.scala
 //
 class Max2 extends Module {
   val io = IO(new Bundle {
