+++ src/main/scala/problems/MaxN.scala
@@ -6,7 +6,7 @@ import chisel3._
 // Problem:
 //
 // Implement test for this module. Please edit:
-// $chisel-tutorial/src/test/scala/problems/MaxN.scala
+// .../chisel-tutorial/src/test/scala/problems/MaxN.scala
 //
 class MaxN(val n: Int, val w: Int) extends Module {
 
@@ -16,6 +16,5 @@ class MaxN(val n: Int, val w: Int) extends Module {
     val ins = Input(Vec(n, UInt(w.W)))
     val out = Output(UInt(w.W))
   })
-
   io.out := io.ins.reduceLeft(Max2)
 }
