+++ src/main/scala/problems/Mul.scala
@@ -6,8 +6,7 @@ import scala.collection.mutable.ArrayBuffer
 
 // Problem:
 //
-// Implement 'z' = 'x' * 'y' by
-// building filling out muls
+// Implement a four-by-four multiplier using a look-up table.
 //
 class Mul extends Module {
   val io = IO(new Bundle {
@@ -15,7 +14,9 @@ class Mul extends Module {
     val y   = Input(UInt(4.W))
     val z   = Output(UInt(8.W))
   })
-  val muls = new ArrayBuffer[UInt]()
+  val mulsValues = new ArrayBuffer[UInt]()
+
+  // Calculate io.z = io.x * io.y by generating a table of values for mulsValues
 
   // Implement below ----------
