+++ src/main/scala/problems/VecShiftRegisterParam.scala
@@ -5,16 +5,17 @@ import chisel3._
 
 // Problem:
 //
-// Implement parametrized simple shift register
-// 'n' is number of elements in shift register
-// 'w' is width of one element
+// Implement a parametrized simple shift register.
+// 'n' is the number of elements in the shift register.
+// 'w' is the width of one element.
 
-// Implement below ----------
 class VecShiftRegisterParam(val n: Int, val w: Int) extends Module {
   val io = IO(new Bundle {
     val in  = Input(UInt(w.W))
     val out = Output(UInt(w.W))
   })
+
+  // Implement below ----------
 
   io.out := 0.U
 }
