+++ src/main/scala/problems/VendingMachine.scala
@@ -6,10 +6,10 @@ import chisel3.util.Enum
 
 // Problem:
 //
-// Implement vending machine using 'when' states
-// 'nickel' is 5 worth coin
-// 'dime'   is 10 worth coin
-// 'sOk' is reached when there are 20 or more worth coins in machine
+// Implement a vending machine using 'when' states.
+// 'nickel' is a 5 cent coin
+// 'dime'   is 10 cent coin
+// 'sOk' is reached when there are coins totalling 20 cents or more in the machine.
 //
 class VendingMachine extends Module {
   val io = IO(new Bundle {
@@ -19,7 +19,7 @@ class VendingMachine extends Module {
   })
   val sIdle :: s5 :: s10 :: s15 :: sOk :: Nil =
     Enum(5)
-  val state = Reg(init=sIdle)
+  val state = Reg(init = sIdle)
 
   // Implement below ----------
