+++ src/main/scala/problems/VendingMachineSwitch.scala
@@ -6,10 +6,10 @@ import chisel3.util._
 
 // Problem:
 //
-// Implement vending machine FSM using 'switch' statement
-// 'nickel' is 5  worth coin
-// 'dime'   is 10 worth coin
-// 's_ok' is reached when there are 20 or more worth coins in machine
+// Implement a vending machine using a 'switch' statement.
+// 'nickel' is a 5 cent coin
+// 'dime'   is 10 cent coin
+// 'sOk' is reached when there are coins totalling 20 cents or more in the machine.
 //
 class VendingMachineSwitch extends Module {
   val io = IO(new Bundle {
@@ -19,7 +19,7 @@ class VendingMachineSwitch extends Module {
   })
   val s_idle :: s_5 :: s_10 :: s_15 :: s_ok :: Nil = Enum(5)
   val state = Reg(init = s_idle)
-  
+
   switch (state) {
     is (s_idle) {
       when (io.nickel) { state := s_5 }
