+++ src/main/scala/solutions/Accumulator.scala
@@ -3,6 +3,11 @@ package solutions
 
 import chisel3._
 
+// Problem:
+//
+// Count incoming trues
+// (increase counter every clock if 'in' is asserted)
+//
 class Accumulator extends Module {
   val io = IO(new Bundle {
     val in  = Input(UInt(1.W))
