+++ src/main/scala/solutions/Counter.scala
@@ -4,23 +4,21 @@ package solutions
 import chisel3._
 // import Counter._
 
+// Problem:
+//
+// Counter should be incremented by the 'amt'
+// every clock if 'en' is asserted
+//
 object Counter {
 
   def wrapAround(n: UInt, max: UInt) = 
     Mux(n > max, 0.U, n)
-
-  // ---------------------------------------- \\
-  // Modify this function to increment by the
-  // amt only when en is asserted
-  // ---------------------------------------- \\
 
   def counter(max: UInt, en: Bool, amt: UInt): UInt = {
     val x = Reg(init=0.asUInt(max.getWidth.W))
     when (en) { x := wrapAround(x + amt, max) }
     x
   }
-
-  // ---------------------------------------- \\
 
 }
