+++ src/main/scala/solutions/LFSR16.scala
@@ -4,6 +4,12 @@ package solutions
 import chisel3._
 import chisel3.util.Cat
 
+// Problem:
+//
+// Implement a 16-bit Fibonacci Linear-feedback shift register
+// with polynomial x^16 + x^14 + x^13 + x^11 + 1
+// State change is allowed only when 'inc' is asserted
+//
 class LFSR16 extends Module {
   val io = IO(new Bundle {
     val inc = Input(Bool())
