+++ src/main/scala/solutions/Mux4.scala
@@ -3,6 +3,26 @@ package solutions
 
 import chisel3._
 
+// Example:
+//
+// This is example of multiplexer 2-to-1 with 'sel' as control signal
+// Multiplexed inputs are 'in0' and 'in1'
+//
+class Mux2 extends Module {
+  val io = IO(new Bundle {
+    val sel = Input(UInt(1.W))
+    val in0 = Input(UInt(1.W))
+    val in1 = Input(UInt(1.W))
+    val out = Output(UInt(1.W))
+  })
+  io.out := (io.sel & io.in1) | (~io.sel & io.in0)
+}
+
+// Problem:
+//
+// Build a 4-to-1 multiplexer out of three 2-to-1 multiplexers
+// The first multiplexer is already done for you
+//
 class Mux4 extends Module {
   val io = IO(new Bundle {
     val in0 = Input(UInt(1.W))
@@ -12,14 +32,6 @@ class Mux4 extends Module {
     val sel = Input(UInt(2.W))
     val out = Output(UInt(1.W))
   })
-
-  //-------------------------------------------------------------------------\\
-
-  // Modify this section to build a 4-to-1 mux out of 3 2-to-1 mux
-  // The first mux is already done for you
-
-
-  //-------------------------------------------------------------------------\\
 
   val m0 = Module(new Mux2())
   m0.io.sel := io.sel(0)
