+++ src/main/scala/solutions/RealGCD.scala
@@ -4,6 +4,10 @@ package solutions
 import chisel3._
 import chisel3.util.{Valid, DeqIO}
 
+// Problem:
+// Implement a GCD circuit (the greatest common divisor of two numbers).
+// Input numbers are bundled as 'RealGCDInput' and communicated using the Decoupled handshake protocol
+//
 class RealGCDInput extends Bundle {
   val a = UInt(16.W)
   val b = UInt(16.W)
