+++ src/main/scala/solutions/VecShiftRegister.scala
@@ -3,6 +3,13 @@ package solutions
 
 import chisel3._
 
+// Problem:
+//
+// Implement a loadable shift register with four 4-bit stages using Vec
+// Shift occurs if 'shift' is asserted
+// Load  occurs if 'load'  is asserted
+// Whole state should be replaced with 'ins' when loaded
+//
 class VecShiftRegister extends Module {
   val io = IO(new Bundle {
     val ins   = Input(Vec(4, UInt(4.W)))
