+++ src/main/scala/solutions/VecShiftRegisterParam.scala
@@ -3,6 +3,12 @@ package solutions
 
 import chisel3._
 
+// Problem:
+//
+// Implement a parametrized simple shift register.
+// 'n' is the number of elements in the shift register.
+// 'w' is the width of one element.
+
 class VecShiftRegisterParam(val n: Int, val w: Int) extends Module {
   val io = IO(new Bundle {
     val in  = Input(UInt(w.W))
