+++ src/main/scala/solutions/VecShiftRegisterSimple.scala
@@ -3,6 +3,11 @@ package solutions
 
 import chisel3._
 
+// Problem:
+//
+// Implement a shift register with four 8-bit stages.
+// Shift should occur on every clock.
+//
 class VecShiftRegisterSimple extends Module {
   val io = IO(new Bundle {
     val in  = Input(UInt(8.W))
