+++ src/test/scala/solutions/Launcher.scala
@@ -62,11 +62,6 @@ object Launcher {
         (c) => new RealGCDTests(c)
       }
     },
-    "Mux2" -> { (backendName: String) =>
-      Driver(() => new Mux2(), backendName) {
-        (c) => new Mux2Tests(c)
-      }
-    },
     "Mux4" -> { (backendName: String) =>
       Driver(() => new Mux4(), backendName) {
         (c) => new Mux4Tests(c)
